requirements.txt
# 深度学习框架
torch>=2.0.0

# 数据处理和分析
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
meteostat>=1.6.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 配置和工具
pyyaml>=6.0.0
tqdm>=4.65.0
joblib>=1.3.0
pathlib>=1.0.1